pyproject.toml
[project]
name = "airas-experiment"
version = "0.1.0"
description = "AIRAS ML Experiment: Calibrated Metacognitive Self-Consistency"
requires-python = ">=3.11"

dependencies = [
    # Core framework
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    
    # Experiment tracking
    "wandb>=0.16.0",
    
    # Data and ML
    "numpy>=1.24.0",
    "datasets>=2.14.0",
    
    # LLM API
    "openai>=1.0.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    
    # Optional dependencies
    "scikit-learn>=1.3.0",  # For isotonic regression calibration
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"